% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{distance_correlation}
\alias{distance_correlation}
\title{Compute the correlation between pairwise distances}
\usage{
distance_correlation(pseudotime, truetime)
}
\arguments{
\item{pseudotime}{An n-vector of pseudotime estimates.}

\item{truetime}{An n-vector of true time values.}
}
\value{
The spearman correlation between computed distances.
}
\description{
This computes the pairwise distances between cells according to the pseudotime
estimates and truetime points, and then returns the Spearman correlation between these
distances.
It is recommended to normalize the pseudotime and truetime values before calling this function.
}
