% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{flip}
\alias{flip}
\title{Flips the pseudotime estimates over their midpoint}
\usage{
flip(pseudotime)
}
\arguments{
\item{pseudotime}{A vector of pseudotime estimates}
}
\value{
A vector of flipped pseudotime estimates
}
\description{
Flips the pseudotime estimates over their midpoint
}
