% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{lasso_coef_correlation}
\alias{lasso_coef_correlation}
\title{Correlation between LASSO coefficients.}
\usage{
lasso_coef_correlation(
  expression_matrix,
  pseudotime,
  truetime,
  milestones,
  ...
)
}
\arguments{
\item{expression_matrix}{A p x n normalized gene expression matrix.}

\item{pseudotime}{An n-vector of pseudotime estimates.}

\item{truetime}{An n-vector of true timpoints.}

\item{milestones}{A subset of the true time points, used to compute LASSO coefficients.}

\item{...}{Additional parameters passed to \code{cv.glmnet}.}
}
\description{
Similar to the random forest correlation metric, but rather than Gini impurity indices we compute the coefficients corresponding to each gene with a LASSO regression model.
}
