% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mae}
\alias{mae}
\title{Compute the mean absolute error between pseudotime estimates and true timepoints.}
\usage{
mae(pseudotime, truetime)
}
\arguments{
\item{pseudotime}{An n-vector of pseudotime estimates.}

\item{truetime}{An n-vector of true timepoint values.}
}
\value{
The MAE between pseudotime and truetime.
}
\description{
Compute the mean absolute error between pseudotime estimates and true timepoints.
}
