% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Compute the mean squared error between pseudotime estimates and true timepoints.}
\usage{
mse(pseudotime, truetime)
}
\arguments{
\item{pseudotime}{An n-vector of pseudotime estimates.}

\item{truetime}{An n-vector of true timepoint values.}
}
\value{
The MSE between pseudotime and truetime.
}
\description{
Compute the mean squared error between pseudotime estimates and true timepoints.
}
