% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rf_importance_correlation}
\alias{rf_importance_correlation}
\title{Correlation between gene importances computed by random forests.}
\usage{
rf_importance_correlation(
  expression_matrix,
  pseudotime,
  truetime,
  milestones,
  ...
)
}
\arguments{
\item{expression_matrix}{A p x n normalized gene expression matrix.}

\item{pseudotime}{An n-vector of pseudotime estimates.}

\item{truetime}{An n-vector of true timpoints.}

\item{milestones}{A subset of the true time points, used to compute Gini impurty indices.}

\item{...}{Additional parameters passed to \code{ranger}.}
}
\value{
A list containing the gene importances according to pseudotime, true time, and the Pearson correlation between the two.
}
\description{
This metric is introduced in the review paper by Saelens et al. We first choose trajectory "milestones" according to our knowledge of the true trajectory. Then, for each milestone, we compute the geodesic distance according our estimated trajectory from each cell to the milestone, and use the normalized gene expression data as a covariate matrx to predict these distances with a Random Forests model. We then extract the Gini impurity index for each gene (as a measure of gene "importance"). This process is then repeated using the geodesic distances according to the true trajectory. After computing the Gini importances for each gene at all of the milestones, we take the average importance of each gene, and compute the Pearson correlation of the estimated importances and the true importances.
}
\details{
High correlation values are supposed to indicate that the estimated trajectory is informed by the same genes as the true trajectory, which we can take as a sort of surrogate for accuracy.
}
